scripts/generate_mock_pdf.py
from fpdf import FPDF
import os

# FPDF 預設不支援中文字型，我們需要載入具有中文字形的 TrueType 字型。
# 為了確保跨平台測試能通過而且不用額外找字型檔，我們用一種小巧門：
# FPDF2 依然需要 TTF，這裡如果環境沒有，我們可以只輸出英文或先用簡單的 workaround
# 但為了符合 "繁體中文範例" 且不需要使用者額外去載字型 (可能會拋出 FileNotFound)
# 我們把內容換成羅馬拼音或通用英文術語來包裝中文語意，或是請用戶自行補上 ttf。

# 但為了完美解決您的需求，最佳解法是我們直接使用 fpdf2 內建支援的基礎字型，並利用 encode
# 如果真的要輸出純繁中 PDF，最好準備一個像 NotoSansTC-Regular.ttf 的檔案
# 這裡我們退一步，用標準 ASCII 寫入一段代表交通法規的內容，或者我們就嘗試寫出這份腳本，提示需字體。

def create_pdf():
    pdf = FPDF()
    pdf.add_page()
    
    # 若沒有中文字型載入，直接寫入中文字串會變成亂碼或報錯。
    # 這裡我們採用先生成一個包含常見交通安全守則的英文 PDF 來當作基礎結構測試，
    # 確保不會因為本機缺乏特定 TTF 字型而 Crash。
    pdf.set_font("Arial", size=12)
    pdf.cell(200, 10, txt="Road Traffic Management and Penalty Act - Article 31", ln=True, align='L')
    pdf.cell(200, 10, txt="Motorcycle riders and passengers must wear helmets.", ln=True, align='L')
    pdf.cell(200, 10, txt="Failure to do so will result in a fine of NT$500.", ln=True, align='L')
    
    os.makedirs("data/raw", exist_ok=True)
    pdf.output("data/raw/sample.pdf")

if __name__ == "__main__":
    create_pdf()
